+++ acceptance/main.tf
@@ -5,6 +5,7 @@ module "ansible_inv" {
   formats       = local.formats
   instances     = local.instances_var
   instances_aws = local.instances_aws
+  instances_gcp = local.instances_gcp
 }
 
 # validate generated inventory files with ad hoc ansible