+++ locals.tf
@@ -23,6 +23,6 @@ locals {
 
   # transform gcp instances object into expected structure
   instances_gcp_transform = {
-    for instance in var.instances_gcp : try(instance.tags[0], instance.instance_id) => merge({ "ansible_host" = instance.network_interface.0.network_ip }, { for idx, tag in instance.tags : "var${idx}" => tag })
+    for instance in var.instances_gcp : try(instance.tags[0], instance.instance_id) => merge({ "ansible_host" = instance.network_interface.0.network_ip }, { for tag in instance.tags : regexall("[-\\w]+", tag)[0] => regexall("[-\\w]+", tag)[1] if length(regexall("[-\\w]+", tag)) == 2 })
   }
 }