+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.0.1 (Next)
 - Support direct input of AWS instance resource objects.
+- Support direct input of GCP instance resource objects.
 
 ### 1.0.0
 - Initial Release