+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-### 1.0.1 (Next)
+### 1.0.2 (Next)
+
+### 1.0.1
 - Support direct input of AWS instance resource objects.
 - Support direct input of GCP instance resource objects.
 - Support direct input of Azure instance resource objects.